requirements.txt
Flask==3.0.3
flask-socketio==5.3.7
eventlet==0.36.1
python-dotenv==1.0.1
websocket-client==1.8.0
Pillow==10.4.0
pytesseract==0.3.13
